Extract a server start method into the server lib
